email: Move `TokenExposedEmail` struct to its corresponding module
