+++ src/controllers/github/secret_scanning.rs
@@ -1,6 +1,6 @@
 use crate::app::AppState;
 use crate::controllers::frontend_prelude::*;
-use crate::email::TokenExposedEmail;
+use crate::email::Email;
 use crate::models::{ApiToken, User};
 use crate::schema::api_tokens;
 use crate::util::token::HashedToken;
@@ -192,6 +192,40 @@ fn send_notification_email(
     state.emails.send(&recipient, email)?;
 
     Ok(())
+}
+
+struct TokenExposedEmail<'a> {
+    domain: &'a str,
+    reporter: &'a str,
+    source: &'a str,
+    token_name: &'a str,
+    url: &'a str,
+}
+
+impl Email for TokenExposedEmail<'_> {
+    const SUBJECT: &'static str = "Exposed API token found";
+
+    fn body(&self) -> String {
+        let mut body = format!(
+            "{reporter} has notified us that your crates.io API token {token_name}\n
+has been exposed publicly. We have revoked this token as a precaution.\n
+Please review your account at https://{domain} to confirm that no\n
+unexpected changes have been made to your settings or crates.\n
+\n
+Source type: {source}\n",
+            domain = self.domain,
+            reporter = self.reporter,
+            source = self.source,
+            token_name = self.token_name,
+        );
+        if self.url.is_empty() {
+            body.push_str("\nWe were not informed of the URL where the token was found.\n");
+        } else {
+            body.push_str(&format!("\nURL where the token was found: {}\n", self.url));
+        }
+
+        body
+    }
 }
 
 #[derive(Deserialize, Serialize)]
